+++ src/bunit.web/TestDoubles/NavigationManager/FakeNavigationManager.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using Bunit.Rendering;
 using Microsoft.AspNetCore.Components;
@@ -12,6 +13,16 @@ namespace Bunit.TestDoubles
 	public sealed class FakeNavigationManager : NavigationManager
 	{
 		private readonly ITestRenderer renderer;
+		private readonly Stack<NavigationHistory> history = new();
+
+		/// <summary>
+		/// The navigation history captured by the <see cref="FakeNavigationManager"/>.
+		/// This is a stack based collection, so the first element is the latest/current navigation target.
+		/// </summary>
+		/// <remarks>
+		/// The initial Uri is not added to the history.
+		/// </remarks>
+		public IReadOnlyCollection<NavigationHistory> History => history;
 
 		/// <summary>
 		/// Initializes a new instance of the <see cref="FakeNavigationManager"/> class.
@@ -23,10 +34,12 @@ namespace Bunit.TestDoubles
 			Initialize("http://localhost/", "http://localhost/");
 		}
 
+#if !NET6_0_OR_GREATER
 		/// <inheritdoc/>
 		protected override void NavigateToCore(string uri, bool forceLoad)
 		{
 			Uri = ToAbsoluteUri(uri).OriginalString;
+			history.Push(new NavigationHistory(uri, new NavigationOptions(forceLoad)));
 
 			renderer.Dispatcher.InvokeAsync(() =>
 			{
@@ -34,5 +47,26 @@ namespace Bunit.TestDoubles
 				NotifyLocationChanged(isInterceptedLink: false);
 			});
 		}
+#endif
+
+#if NET6_0_OR_GREATER
+
+		/// <inheritdoc/>
+		protected override void NavigateToCore(string uri, NavigationOptions options)
+		{
+			Uri = ToAbsoluteUri(uri).OriginalString;
+
+			if (options.ReplaceHistoryEntry && history.Count > 0)
+				history.Pop();
+
+			history.Push(new NavigationHistory(uri, options));
+
+			renderer.Dispatcher.InvokeAsync(() =>
+			{
+				Uri = ToAbsoluteUri(uri).OriginalString;
+				NotifyLocationChanged(isInterceptedLink: false);
+			});
+		}	
+#endif
 	}
 }